Renamed list to pspotslist.
