+++ app/parkit/App.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { StyleSheet, Text, View } from "react-native";
 import data from "./assets/temp.json";
 import { IParkingSpot } from "./types/ParkingSpots.js";
-import List from "./use-cases/main/list/List";
+import PSpotsList from "./use-cases/main/list-pspots/PSpotsList";
 
 export default class App extends React.Component {
     private spots: Array<IParkingSpot>;
@@ -17,7 +17,7 @@ export default class App extends React.Component {
         return (
             <View style={styles.container}>
                 <Text>Open up App.tsx to start working on your app!</Text>
-                <List array={this.spots} />
+                <PSpotsList array={this.spots} />
             </View>
         );
     }
