+++ app/parkit/package.json
@@ -59,5 +59,5 @@
     "android": {
       "package": "chalmers.dat-256-weedle.parkit"
     }
-   }
+  }
 }
